% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{all_tokenized}
\alias{all_tokenized}
\alias{all_tokenized_predictors}
\title{Role Selection}
\usage{
all_tokenized()

all_tokenized_predictors()
}
\description{
\code{all_tokenized()} selects all tokenlist columns, \code{all_tokenized_predictors()}
selects all predictor tokenlist columns.
}
\seealso{
\code{\link[recipes:has_role]{recipes::has_role()}}
}
