% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoji_samples}
\alias{emoji_samples}
\title{Sample sentences with emojis}
\format{
tibble with 1 column
}
\usage{
emoji_samples
}
\description{
This data set is primarily used for examples.
}
\keyword{datasets}
