% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{okc_text}
\alias{okc_text}
\title{OkCupid Text Data}
\source{
Kim, A. Y., and A. Escobedo-Land. 2015. "OkCupid Data for
  Introductory Statistics and Data Science Courses." *Journal of
  Statistics Education: An International Journal on the Teaching and
  Learning of Statistics*.
}
\value{
\item{okc_text}{a tibble}
}
\description{
These are a sample of columns and users of OkCupid dating
 website. The data are from Kim and Escobedo-Land (2015).
 Permission to use this data set was explicitly granted by
 OkCupid. The data set contains 10 text fields filled out by
 users.
}
\examples{
data(okc_text)
str(okc_text)
}
\keyword{datasets}
