% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ngram}
\alias{rcpp_ngram}
\title{ngram generator}
\usage{
rcpp_ngram(x, n, n_min, delim)
}
\arguments{
\item{x}{list of character vectors}

\item{n}{number of grams}

\item{n_min}{minimum number of grams}

\item{delim}{delimiter}
}
\description{
ngram generator
}
