% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashing.R, R/lda.R, R/text_normalization.R,
%   R/textfeature.R, R/tokenize.R
\name{required_pkgs.step_texthash}
\alias{required_pkgs.step_texthash}
\alias{required_pkgs.step_lda}
\alias{required_pkgs.step_text_normalization}
\alias{required_pkgs.step_textfeature}
\alias{required_pkgs.step_tokenize}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
required_pkgs.step_texthash(x, ...)

required_pkgs.step_lda(x, ...)

required_pkgs.step_text_normalization(x, ...)

required_pkgs.step_textfeature(x, ...)

required_pkgs.step_tokenize(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
S3 methods for tracking which additional packages are needed for steps.
}
\keyword{internal}
