% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfeature.R
\name{step_textfeature}
\alias{step_textfeature}
\alias{tidy.step_textfeature}
\title{Calculate Set of Text Features}
\usage{
step_textfeature(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  extract_functions = count_functions,
  prefix = "textfeature",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("textfeature")
)
}
\arguments{
\item{recipe}{A \code{\link[recipes:recipe]{recipes::recipe()}} object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from the
original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{columns}{A character vector of the selected variable names. This is
\code{NULL} until the step is trained by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}}, and should
not be set by hand.}

\item{extract_functions}{A named list of feature extracting functions. Each
function should take a character vector and return a numeric vector of the
same length. Defaults to \code{count_functions}. See details for more
information.}

\item{prefix}{A character string that will be the prefix to the resulting
new variables. See notes below.}

\item{keep_original_cols}{A logical to keep the original variables in the
output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked
by \code{\link[recipes:bake]{recipes::bake()}}? While all operations are baked when
\code{\link[recipes:prep]{recipes::prep()}} is run, some operations may not be able to be conducted
on new data (e.g. processing the outcome variable(s)). Care should be
taken when using \code{skip = TRUE}, as it may affect the computations for
subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added
to the sequence of existing steps (if any).
}
\description{
\code{step_textfeature()} creates a \emph{specification} of a recipe step that will
extract a number of numeric features of a text column.
}
\details{
This step will take a character column and returns a number of numeric
columns equal to the number of functions in the list passed to the
\code{extract_functions} argument.

All the functions passed to \code{extract_functions} must take a character vector
as input and return a numeric vector of the same length, otherwise an error
will be thrown.

The new columns will have names that begin with \code{prefix}, then the name of
the variable, then the name of the function in \code{extract_functions}, all
separated by \verb{_}. For example, with the default \code{prefix = "textfeature"}, a
variable named \code{text}, and the default \code{count_functions}, the columns will
include \code{textfeature_text_n_words} and \code{textfeature_text_n_uq_words}.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms}, \code{functions}, and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{functions}{character, name of feature functions}
\item{id}{character, id of this step}
}
}

\section{Case weights}{


The underlying operation does not allow for case weights.
}

\examples{
\dontshow{if (rlang::is_installed("modeldata")) withAutoprint(\{ # examplesIf}
library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) |>
  step_textfeature(medium)

tate_obj <- tate_rec |>
  prep()

bake(tate_obj, new_data = NULL) |>
  slice(1:2)

bake(tate_obj, new_data = NULL) |>
  pull(textfeature_medium_n_words)

tidy(tate_rec, number = 1)
tidy(tate_obj, number = 1)

# Using custom extraction functions
nchar_round_10 <- function(x) round(nchar(x) / 10) * 10

recipe(~., data = tate_text) |>
  step_textfeature(medium,
    extract_functions = list(nchar10 = nchar_round_10)
  ) |>
  prep() |>
  bake(new_data = NULL)
\dontshow{\}) # examplesIf}
}
\seealso{
Other Steps for Numeric Variables From Characters:
\code{\link[=step_dummy_hash]{step_dummy_hash()}},
\code{\link[=step_sequence_onehot]{step_sequence_onehot()}}
}
\concept{Steps for Numeric Variables From Characters}
