% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenmerge.R
\name{step_tokenmerge}
\alias{step_tokenmerge}
\alias{tidy.step_tokenmerge}
\title{Combine Multiple Token Variables Into One}
\usage{
step_tokenmerge(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  prefix = "tokenmerge",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("tokenmerge")
)
}
\arguments{
\item{recipe}{A \code{\link[recipes:recipe]{recipes::recipe()}} object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from the
original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{columns}{A character vector of the selected variable names. This is
\code{NULL} until the step is trained by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}}, and should
not be set by hand.}

\item{prefix}{A character string that will be the prefix to the resulting
new variables. See notes below.}

\item{keep_original_cols}{A logical to keep the original variables in the
output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked
by \code{\link[recipes:bake]{recipes::bake()}}? While all operations are baked when
\code{\link[recipes:prep]{recipes::prep()}} is run, some operations may not be able to be conducted
on new data (e.g. processing the outcome variable(s)). Care should be
taken when using \code{skip = TRUE}, as it may affect the computations for
subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added
to the sequence of existing steps (if any).
}
\description{
\code{step_tokenmerge()} creates a \emph{specification} of a recipe step that will take
multiple \code{\link[=tokenlist]{token}} variables and combine them into one
\code{\link[=tokenlist]{token}} variable.
}
\details{
All the selected variables are merged into a single new variable named
\code{prefix}, which is \code{"tokenmerge"} by default. The name of the selected
variables is not part of the new name.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms} and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{id}{character, id of this step}
}
}

\section{Case weights}{


The underlying operation does not allow for case weights.
}

\examples{
\dontshow{if (rlang::is_installed("modeldata")) withAutoprint(\{ # examplesIf}
library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) |>
  step_tokenize(medium, artist) |>
  step_tokenmerge(medium, artist)

tate_obj <- tate_rec |>
  prep()

bake(tate_obj, new_data = NULL)

tidy(tate_rec, number = 2)
tidy(tate_obj, number = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=step_tokenize]{step_tokenize()}} to turn characters into \code{\link[=tokenlist]{tokens}}

Other Steps for Token Modification:
\code{\link[=step_lemma]{step_lemma()}},
\code{\link[=step_ngram]{step_ngram()}},
\code{\link[=step_pos_filter]{step_pos_filter()}},
\code{\link[=step_stem]{step_stem()}},
\code{\link[=step_stopwords]{step_stopwords()}},
\code{\link[=step_tokenfilter]{step_tokenfilter()}}
}
\concept{Steps for Token Modification}
