% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textrecipes-package.R
\docType{package}
\name{textrecipes-package}
\alias{textrecipes}
\alias{textrecipes-package}
\title{textrecipes: Extra 'Recipes' for Text Processing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Converting text to numerical features requires specifically created procedures, which are implemented as steps according to the 'recipes' package. These steps allows for tokenization, filtering, counting (tf and tfidf) and feature hashing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/textrecipes}
  \item \url{https://textrecipes.tidymodels.org/}
  \item Report bugs at \url{https://github.com/tidymodels/textrecipes/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})
}

Other contributors:
\itemize{
  \item Michael W. Kearney \email{kearneymw@missouri.edu} (author of count_functions) [copyright holder]
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
