% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_levels.R, R/clean_names.R,
%   R/dummy_hash.R, R/lda.R, R/lemma.R, R/ngram.R, R/pos_filter.R,
%   R/sequence_onehot.R, R/stem.R, R/stopwords.R, R/text_normalization.R,
%   R/textfeature.R, R/texthash.R, R/tf.R, R/tfidf.R, R/tidy.R,
%   R/tokenfilter.R, R/tokenize.R, R/tokenize_bpe.R, R/tokenize_sentencepiece.R,
%   R/tokenize_wordpiece.R, R/tokenmerge.R, R/untokenize.R, R/word_embeddings.R
\name{tidy.step_clean_levels}
\alias{tidy.step_clean_levels}
\alias{tidy.step_clean_names}
\alias{tidy.step_dummy_hash}
\alias{tidy.step_lda}
\alias{tidy.step_lemma}
\alias{tidy.step_ngram}
\alias{tidy.step_pos_filter}
\alias{tidy.step_sequence_onehot}
\alias{tidy.step_stem}
\alias{tidy.step_stopwords}
\alias{tidy.step_text_normalization}
\alias{tidy.step_textfeature}
\alias{tidy.step_texthash}
\alias{tidy.step_tf}
\alias{tidy.step_tfidf}
\alias{tidy.recipe}
\alias{tidy.step_tokenfilter}
\alias{tidy.step_tokenize}
\alias{tidy.step_tokenize_bpe}
\alias{tidy.step_tokenize_sentencepiece}
\alias{tidy.step_tokenize_wordpiece}
\alias{tidy.step_tokenmerge}
\alias{tidy.step_untokenize}
\alias{tidy.step_word_embeddings}
\title{Tidy the Result of a Recipe}
\usage{
\method{tidy}{step_clean_levels}(x, ...)

\method{tidy}{step_clean_names}(x, ...)

\method{tidy}{step_dummy_hash}(x, ...)

\method{tidy}{step_lda}(x, ...)

\method{tidy}{step_lemma}(x, ...)

\method{tidy}{step_ngram}(x, ...)

\method{tidy}{step_pos_filter}(x, ...)

\method{tidy}{step_sequence_onehot}(x, ...)

\method{tidy}{step_stem}(x, ...)

\method{tidy}{step_stopwords}(x, ...)

\method{tidy}{step_text_normalization}(x, ...)

\method{tidy}{step_textfeature}(x, ...)

\method{tidy}{step_texthash}(x, ...)

\method{tidy}{step_tf}(x, ...)

\method{tidy}{step_tfidf}(x, ...)

\method{tidy}{step_tokenfilter}(x, ...)

\method{tidy}{step_tokenize}(x, ...)

\method{tidy}{step_tokenize_bpe}(x, ...)

\method{tidy}{step_tokenize_sentencepiece}(x, ...)

\method{tidy}{step_tokenize_wordpiece}(x, ...)

\method{tidy}{step_tokenmerge}(x, ...)

\method{tidy}{step_untokenize}(x, ...)

\method{tidy}{step_word_embeddings}(x, ...)
}
\arguments{
\item{x}{A \code{step_word_embeddings} object.}

\item{...}{Not currently used.}
}
\description{
\code{tidy} will return a data frame that contains information regarding a recipe
or operation within the recipe (when a \code{tidy} method for the operation
exists). See \link[recipes:tidy.recipe]{recipes::tidy.recipe} for more information.
}
