% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashing.R, R/hashing_dummy.R, R/ngram.R,
%   R/tf.R, R/tokenfilter.R, R/tokenize.R
\name{tunable.step_texthash}
\alias{tunable.step_texthash}
\alias{tunable.step_dummy_hash}
\alias{tunable.step_ngram}
\alias{tunable.step_tf}
\alias{tunable.step_tokenfilter}
\alias{tunable.step_tokenize}
\title{Find recommended methods for generating parameter values}
\usage{
\method{tunable}{step_texthash}(x, ...)

\method{tunable}{step_dummy_hash}(x, ...)

\method{tunable}{step_ngram}(x, ...)

\method{tunable}{step_tf}(x, ...)

\method{tunable}{step_tokenfilter}(x, ...)

\method{tunable}{step_tokenize}(x, ...)
}
\arguments{
\item{x}{A recipe step}

\item{...}{Not currently used.}
}
\description{
\code{\link[=tunable]{tunable()}} determines which parameters in an object \emph{can} be tuned along
with information about the parameters.
}
\keyword{internal}
