% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{tunable.step_ngram}
\alias{tunable.step_ngram}
\title{tunable methods for step_ngram}
\usage{
tunable.step_ngram(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can
be used to generate new parameter values and other characteristics.
}
\keyword{internal}
