% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_hash.R, R/ngram.R, R/texthash.R, R/tf.R,
%   R/tokenfilter.R, R/tokenize.R, R/tokenize_bpe.R, R/tunable.R
\name{tunable.step_dummy_hash}
\alias{tunable.step_dummy_hash}
\alias{tunable.step_ngram}
\alias{tunable.step_texthash}
\alias{tunable.step_tf}
\alias{tunable.step_tokenfilter}
\alias{tunable.step_tokenize}
\alias{tunable.step_tokenize_bpe}
\alias{tunable_textrecipes}
\title{tunable methods for textrecipes}
\usage{
\method{tunable}{step_dummy_hash}(x, ...)

\method{tunable}{step_ngram}(x, ...)

\method{tunable}{step_texthash}(x, ...)

\method{tunable}{step_tf}(x, ...)

\method{tunable}{step_tokenfilter}(x, ...)

\method{tunable}{step_tokenize}(x, ...)

\method{tunable}{step_tokenize_bpe}(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can
be used to generate new parameter values and other characteristics.
}
\keyword{internal}
