% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{all_tokenized}
\alias{all_tokenized}
\alias{all_tokenized_predictors}
\title{Role Selection}
\usage{
all_tokenized()

all_tokenized_predictors()
}
\description{
\code{all_tokenized()} selects all \code{\link[=tokenlist]{token}} variables,
\code{all_tokenized_predictors()} selects all predictor \code{\link[=tokenlist]{token}}
variables.
}
\seealso{
\code{\link[recipes:has_role]{recipes::has_role()}}
}
