% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{step_lda}
\alias{step_lda}
\alias{tidy.step_lda}
\title{Calculate LDA Dimension Estimates of Tokens}
\source{
\url{https://arxiv.org/abs/1301.3781}
}
\usage{
step_lda(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  lda_models = NULL,
  num_topics = 10L,
  prefix = "lda",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("lda")
)
}
\arguments{
\item{recipe}{A \code{\link[recipes:recipe]{recipes::recipe()}} object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from the
original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{columns}{A character vector of the selected variable names. This is
\code{NULL} until the step is trained by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}}, and should
not be set by hand.}

\item{lda_models}{A WarpLDA model object from \link[text2vec:LDA]{text2vec::LDA}. If
\code{NULL}, the default, it will train its model based on the training
data. Look at the examples for how to fit a WarpLDA model.}

\item{num_topics}{A whole number, the desired number of latent topics.
Defaults to \code{10}.}

\item{prefix}{A character string that will be the prefix to the resulting
new variables. See notes below.}

\item{keep_original_cols}{A logical to keep the original variables in the
output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked
by \code{\link[recipes:bake]{recipes::bake()}}? While all operations are baked when
\code{\link[recipes:prep]{recipes::prep()}} is run, some operations may not be able to be conducted
on new data (e.g. processing the outcome variable(s)). Care should be
taken when using \code{skip = TRUE}, as it may affect the computations for
subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added
to the sequence of existing steps (if any).
}
\description{
\code{step_lda()} creates a \emph{specification} of a recipe step that will return the
lda dimension estimates of a text variable.
}
\details{
The new columns will have names that begin with \code{prefix}, then the name of
the variable, then the index of the topic, all separated by \verb{_}. For example,
with the default \code{prefix = "lda"} and a variable named \code{text}, the columns
will be named \code{lda_text_1}, \code{lda_text_2}, and so on up to \code{num_topics}.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms}, \code{num_topics}, and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{num_topics}{integer, number of topics}
\item{id}{character, id of this step}
}
}

\section{Case weights}{


The underlying operation does not allow for case weights.
}

\examples{
\dontshow{if (all(c("modeldata", "text2vec") \%in\% rownames(installed.packages()))) withAutoprint(\{ # examplesIf}
library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) |>
  step_tokenize(medium) |>
  step_lda(medium)

tate_obj <- tate_rec |>
  prep()

bake(tate_obj, new_data = NULL) |>
  slice(1:2)
tidy(tate_rec, number = 2)
tidy(tate_obj, number = 2)

# Changing the number of topics.
recipe(~., data = tate_text) |>
  step_tokenize(medium, artist) |>
  step_lda(medium, artist, num_topics = 20) |>
  prep() |>
  bake(new_data = NULL) |>
  slice(1:2)

# Supplying A pre-trained LDA model trained using text2vec
library(text2vec)
tokens <- word_tokenizer(tolower(tate_text$medium))
it <- itoken(tokens, ids = seq_along(tate_text$medium))
v <- create_vocabulary(it)
dtm <- create_dtm(it, vocab_vectorizer(v))
lda_model <- LDA$new(n_topics = 15)

recipe(~., data = tate_text) |>
  step_tokenize(medium, artist) |>
  step_lda(medium, artist, lda_models = lda_model) |>
  prep() |>
  bake(new_data = NULL) |>
  slice(1:2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other Steps for Numeric Variables From Tokens:
\code{\link[=step_texthash]{step_texthash()}},
\code{\link[=step_tf]{step_tf()}},
\code{\link[=step_tfidf]{step_tfidf()}},
\code{\link[=step_word_embeddings]{step_word_embeddings()}}
}
\concept{Steps for Numeric Variables From Tokens}
