% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{step_ngram}
\alias{step_ngram}
\alias{tidy.step_ngram}
\title{Generate ngrams from tokenlist}
\usage{
step_ngram(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  columns = NULL,
  num_tokens = 3L,
  min_num_tokens = 3L,
  delim = "_",
  skip = FALSE,
  id = rand_id("ngram")
)

\method{tidy}{step_ngram}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables.
For \code{step_ngram}, this indicates the variables to be encoded
into a \link{tokenlist}. See \code{\link[recipes:selections]{recipes::selections()}} for more
details. For the \code{tidy} method, these are not currently used.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the recipe has been
baked.}

\item{columns}{A list of tibble results that define the
encoding. This is \code{NULL} until the step is trained by
\code{\link[recipes:prep]{recipes::prep.recipe()}}.}

\item{num_tokens}{The number of tokens in the n-gram. This must be an integer
greater than or equal to 1. Defaults to 3.}

\item{min_num_tokens}{The minimum number of tokens in the n-gram.
This must be an integer greater than or equal to 1 and smaller than \code{n}.
Defaults to 3.}

\item{delim}{The separator between words in an n-gram. Defaults to "_".}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{recipes::bake.recipe()}}? While all
operations are baked when \code{\link[recipes:prep]{recipes::prep.recipe()}} is run, some
operations may not be able to be conducted on new data (e.g.
processing the outcome variable(s)). Care should be taken when
using \code{skip = TRUE} as it may affect the computations for
subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A \code{step_ngram} object.}
}
\value{
An updated version of \code{recipe} with the new step added
to the sequence of existing steps (if any).
}
\description{
\code{step_ngram} creates a \emph{specification} of a recipe step that
will convert a \link{tokenlist} into a list of ngram of tokens.
}
\details{
The use of this step will leave the ordering of the tokens meaningless.
If \code{min_num_tokens <  num_tokens} then the tokens order in increasing
fashion with respect to the number of tokens in the n-gram. If
\code{min_num_tokens = 1} and \code{num_tokens = 3} then the output contains all the
1-grams followed by all the 2-grams followed by all the 3-grams.
}
\examples{
library(recipes)
library(modeldata)
data(okc_text)

okc_rec <- recipe(~ ., data = okc_text) \%>\%
  step_tokenize(essay0) \%>\%
  step_ngram(essay0)
  
okc_obj <- okc_rec \%>\%
  prep()

juice(okc_obj, essay0) \%>\% 
  slice(1:2)

juice(okc_obj) \%>\% 
  slice(2) \%>\% 
  pull(essay0) 
  
tidy(okc_rec, number = 2)
tidy(okc_obj, number = 2)

}
\seealso{
\code{\link[=step_tokenize]{step_tokenize()}} to turn character into tokenlist.

Other tokenlist to tokenlist steps: 
\code{\link{step_lemma}()},
\code{\link{step_pos_filter}()},
\code{\link{step_stem}()},
\code{\link{step_stopwords}()},
\code{\link{step_tokenfilter}()},
\code{\link{step_tokenmerge}()}
}
\concept{tokenlist to tokenlist steps}
