% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{step_tfidf}
\alias{step_tfidf}
\alias{tidy.step_tfidf}
\title{Term Frequency-Inverse Document Frequency of Tokens}
\usage{
step_tfidf(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  vocabulary = NULL,
  res = NULL,
  smooth_idf = TRUE,
  norm = "l1",
  sublinear_tf = FALSE,
  prefix = "tfidf",
  sparse = "auto",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("tfidf")
)
}
\arguments{
\item{recipe}{A \code{\link[recipes:recipe]{recipes::recipe()}} object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from the
original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{columns}{A character vector of the selected variable names. This is
\code{NULL} until the step is trained by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}}, and should
not be set by hand.}

\item{vocabulary}{A character vector of the tokens to create variables for.
If \code{NULL}, the default, all tokens found in the training data are used.}

\item{res}{A named list of numeric vectors holding the IDF weight of each
token, one element per selected variable. This is \code{NULL} until the step is
trained by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}} and should not be set by hand.}

\item{smooth_idf}{A logical. Should the IDF weights be smoothed by adding one
to the ratio of the number of documents to the number of documents
containing the token? This prevents tokens that appear in every document
from getting a weight of zero. Defaults to \code{TRUE}.}

\item{norm}{A single string determining the type of normalization applied to
the term frequencies of each document. Must be one of:
\itemize{
\item \code{"l1"}: scale by the sum of the term frequencies in the document.
\item \code{"l2"}: scale by the square root of the sum of the squared term
frequencies in the document.
\item \code{"none"}: don't normalize.
}

Defaults to \code{"l1"}.}

\item{sublinear_tf}{A logical. Should sublinear term-frequency scaling be
applied, replacing the term frequency with \code{1 + log(TF)}? Defaults to
\code{FALSE}.}

\item{prefix}{A character string that will be the prefix to the resulting
new variables. See notes below.}

\item{sparse}{A single string. Should the columns produced be sparse
vectors? Can take the values \code{"yes"}, \code{"no"}, and \code{"auto"}. If
\code{sparse = "auto"} then workflows can determine the best option. Defaults
to \code{"auto"}.}

\item{keep_original_cols}{A logical to keep the original variables in the
output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked
by \code{\link[recipes:bake]{recipes::bake()}}? While all operations are baked when
\code{\link[recipes:prep]{recipes::prep()}} is run, some operations may not be able to be conducted
on new data (e.g. processing the outcome variable(s)). Care should be
taken when using \code{skip = TRUE}, as it may affect the computations for
subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added
to the sequence of existing steps (if any).
}
\description{
\code{step_tfidf()} creates a \emph{specification} of a recipe step that will convert a
\code{\link[=tokenlist]{token}} variable into multiple variables containing the term
frequency-inverse document frequency of tokens.
}
\details{
It is strongly advised to use \link{step_tokenfilter} before using \link{step_tfidf} to
limit the number of variables created; otherwise you may run into memory
issues. A good strategy is to start with a low token count and increase
depending on how much RAM you want to use.

Term frequency-inverse document frequency is the product of two statistics:
the term frequency (TF) and the inverse document frequency (IDF).

Term frequency measures how many times each token appears in each
observation.

Inverse document frequency is a measure of how informative a word is, e.g.,
how common or rare the word is across all the observations. If a word appears
in all the observations it might not give that much insight, but if it only
appears in some it might help differentiate between observations.

The IDF is defined as follows when \code{smooth_idf = TRUE}: idf = log(1 + (#
documents in the corpus) / (# documents where the term appears)). When
\code{smooth_idf = FALSE} the \verb{1 +} is dropped.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms}, \code{token}, \code{weight}, and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{token}{character, name of token}
\item{weight}{numeric, the calculated IDF weight}
\item{id}{character, id of this step}
}

The new columns will have names that begin with \code{prefix}, then the name of
the variable, then the token, all separated by \verb{_}. For example, with the
default \code{prefix = "tfidf"}, a variable named \code{text}, and the token \code{"word"},
the new column will be named \code{tfidf_text_word}.
}

\section{Sparse data}{


This step produces sparse columns if \code{sparse = "yes"} is being set. The
default value \code{"auto"} won't trigger production of sparse columns if a recipe
is \code{\link[recipes:prep]{recipes::prep()}}ed, but allows for a workflow to toggle to \code{"yes"} or
\code{"no"} depending on whether the model supports \link[recipes:sparse_data]{recipes::sparse_data} and if
the model is expected to run faster with the data.

The mechanism for determining how much sparsity is produced isn't perfect,
and there will be times when you want to manually overwrite by setting
\code{sparse = "yes"} or \code{sparse = "no"}.
}

\section{Case weights}{


The underlying operation does not allow for case weights.
}

\examples{
\dontshow{if (rlang::is_installed("modeldata")) withAutoprint(\{ # examplesIf}
\donttest{
library(recipes)
library(modeldata)
data(tate_text)

tate_rec <- recipe(~., data = tate_text) |>
  step_tokenize(medium) |>
  step_tfidf(medium)

tate_obj <- tate_rec |>
  prep()

bake(tate_obj, tate_text)

tidy(tate_rec, number = 2)
tidy(tate_obj, number = 2)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=step_tokenize]{step_tokenize()}} to turn characters into \code{\link[=tokenlist]{tokens}}

Other Steps for Numeric Variables From Tokens:
\code{\link[=step_lda]{step_lda()}},
\code{\link[=step_texthash]{step_texthash()}},
\code{\link[=step_tf]{step_tf()}},
\code{\link[=step_word_embeddings]{step_word_embeddings()}}
}
\concept{Steps for Numeric Variables From Tokens}
