% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenlist.R
\name{tokenlist}
\alias{tokenlist}
\title{Create tokenlist object}
\usage{
tokenlist(x = list(), lemma = NULL, pos = NULL)
}
\arguments{
\item{x}{List of character vectors}

\item{lemma}{List of character vectors, must be same size and shape as \code{x}.}

\item{pos}{List of character vectors, must be same size and shape as \code{x}.}
}
\value{
a \link{tokenlist} object.
}
\description{
Create tokenlist object
}
\examples{
tokenlist(list(letters, LETTERS))

tibble(text = tokenlist(list(letters, LETTERS)))
}
